ai-workflow-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.9.0
elasticsearch==8.12.0
python-dotenv==1.0.0
cryptography>=41.0.0
fpdf2==2.7.7
python-docx==1.1.0
requests==2.31.0
motor==3.6.0
pymongo==4.9.1
psutil>=5.9.0
